+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ All notable changes to this project are documented in this file.
 -----------------------
 - Implemented `sendtoaddress` RPC method `#556 <https://github.com/CityOfZion/neo-python/pull/556>`_
 - Gracefully handle network packet deserialization failures
+- Fix issue resetting storage between Smart Contract invocations
 
 [0.7.8] 2018-09-06
 ------------------